business/products/live-seller/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
sqlalchemy==2.0.41
cryptography==44.0.3
pyperclip==1.9.0
httpx==0.28.1
python-escpos==3.1
pyusb==1.3.1
requests==2.32.3
pydantic==2.11.4
pydantic-settings==2.9.1
